Clutch now respects UserDefiend metrics, pins high first.
Clutch will now always scale on a UserDefined metric if one is present
in the metrics stream. It will also now pin high for the first config so
as to prevent underscaling while it is learning about the environment.

Also added some additional unit tests for the control loop.
Finally added license headers.