Add validator for `WorkerPorts`.
Mantis Master uses the legacy constructor. This is now tested and
documented right in the test.

Intent is to add validation to this class so we can validate it from the
Master side when it goes to initialize running workers. If validation
fails, this is considered a precondition which the Master will fail out
the worker. All invalid workers will be resubmitted together after the
launch sequence so they can try to get new port allocations.