Major refactor on providers, discovery, drainer, transmitter.
1. There was a circular dependency between the Configuration and the Job
Discovery. This commit removes that.

2. Additionally, an event drainer performed job discovery (through the
configuration object), to retrieve a Job Cluster using a stream. It then
sent a job cluster as a param to the transmitter. The transmitter should
be responsible for doing this. It already has a job discovery object as
well.

Move the job discovery-related logic from Configuration into Job
Discovery and make an event drainer drain to a stream. The transmitter
is then responsible for implementing how that event is sent to the
stream. In the app-stream flow, the event transmitter will look at its
configuration to get an appName, take that with the stream name, and
call job discovery to determine the right job cluster to send to and
send the event to that destination.

3. This commit also adds in javax.inject providers. Users are free to
use it standalone using the constructors or bring their own DI
framework such as Guice, Spring, etc.