Catch runtime i/o failure on datafile creation.
For S3-backed systems, there's a case where we try to access the file
but it doesn't exist due to read-after-write consistency on S3. This
will throw a runtime exception which will fail out the observable. This
needs to be fixed in the writer, but in the meantime, let's make sure we
have visibility through metrics and logs.