Fix writer file close handling.
The FileAppender#close call can fail if the backing file system fails.
For example, in an S3-backed file system, the call may fail if the call
to GET the file from S3 fails. The file would have been closed so
there's no need to write to it anymore, just create a new file.