Rename defaultStream to __default__.
We experienced a bug in production in which the `mreAppJobClusterMap`
uses __default__ as the routing key when a stream does not have a
routing entry. These routing keys are internally used as stream names,
however internally mantis-published used `defaultStream`. This commit
reconciles these two by renaming defaultStream to __default__.
`
More work to come in the future to avoid conflation of routing keys and
stream names.