[Part 1]: Support for hot task executors in Mantis (#145)
* [Part 1]: Support for hot task executors in Mantis

As we are approaching adding support for Mantis on Titus, one of the first goals is to be able to support task executors that are fully loaded (hot) and capable of executing Mantis stages on demand. This diff is the first step in that direction. This change introduces a new API on the server-side that establishes the protocol that the task executor needs to use to communicate with the server.

Co-authored-by: Sundaram Ananthanarayanan <sananthanarayanan@netflix.com>