Setting preconditions to ensure we don't start mantis with an invalid state.
Currently, we ignore duplicate workers for the same (jobId/stageNum/workerIdx) tuple. However, when we ignore these duplicate workers, we are essentially running a mantis master with a state that may not reflect the previous state of the master. This diff aims to surface these issues and block deployments to make sure we at least know when such problems happen. The current thinking is we should never see these issues. But, if we do see them, then the expectation is that we run state cleaner to clean up the state and then restart the master.