Ensure Master Monitor Can Read Lock ByteBuffer More than Once
It is possible that the DynamoDBMasterMonitor checks for the current
leader more often than the underlying ByteBuffer is updated by
the DynamoDB library.  In this case, we should rewind the buffer.

We can also optionally create a separate config for the lock client
update and the monitor check leader interval to ensure check leader
is called less frequently than the lock client updates.