Handle Case Where Recently Launched Worker Does Not Immediately Heartbeat
It seems like there is a race condition where a recently launched worker has
not sent a heartbeat, the duration is still within the missed heartbeat
threshold and the JobActor treats the lack of heartbeat with a resubmit.
If this is true, this can lead to mass resubmits when a new leader is elected.