Add more efficient mechanism for managing connections
For many types of jobs, connection changes will be far less frequent
then `route` calls.  Therefore, it can be quite expensive to
re-calculate the underlying connection routing on every call to `route`.

This update introduces a new router type `ProactiveRouter`.  These
routers effectively subscribe to changes in the connections and are
responsible for maintaining the state of the connections when routing.
For example on the `ConsistentHashingRouter` this means that they can
just add or remove a connection from the map instead of re-calculating
everything from scratch.

I disabled this feature by default and required it to be explicitly
turned on.  Additionally, I add some hooks to the RouterFactory (should
be the only breaking change) so that we can hook in our own custom
router factories for these connections.