Add ability to set observeOn buffer size per stage
We've found that these buffers are very important tuning parameters.
Additionally, they are very stage specific.  Sometimes, you need want
small buffers and other times large buffers depending on the performance
characteristics of the stage.  Setting them all via the default RX
configuration is not a strong strategy.  Especially because that needs
to be set when mantis-agent spins up; not necessarily when the job spins
up.

This commit is mostly plumbing to get the buffer size to the stage
executors.